+++ modules-template/prov-template/src/main/java/org/openprovenance/prov/template/expander/exception/BundleVariableHasMultipleValues.java
@@ -1,8 +1,7 @@
 package org.openprovenance.prov.template.expander.exception;
 
-import java.util.List;
-
 import org.openprovenance.prov.model.QualifiedName;
+import org.openprovenance.prov.template.json.Descriptors;
 
 public class BundleVariableHasMultipleValues extends RuntimeException {
 
@@ -12,7 +11,7 @@ public class BundleVariableHasMultipleValues extends RuntimeException {
     private static final long serialVersionUID = 1270038830746760223L;
 
     public BundleVariableHasMultipleValues(QualifiedName id,
-	    List<QualifiedName> vals) {
+                                           Descriptors vals) {
     }
 
 }