+++ RELEASES.md
@@ -7,6 +7,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * prov-template: fixed isue with groups
 * prov-template: preparing to replace oldBindings by new Bindings
 * prov-template: WIP: running new Bindings in parallel with OldBindings, and comparing
+* prov-template: relying on new descriptor-based Bindings (no longer using OldBindings and old env structure)
 
 `2.2.2`
 * prov-template: added ptm-merging template